# docs/swagger.yaml
definitions:
  dto.CreateTaskRequest:
    properties:
      headers:
        additionalProperties: true
        type: object
      method:
        type: string
      url:
        type: string
    type: object
  dto.CreateTaskResponse:
    properties:
      id:
        type: integer
    type: object
  dto.GetTaskResponse:
    properties:
      headers:
        additionalProperties: true
        description: 'TODO: headers from 3rd party service response'
        type: object
      httpStatusCode:
        type: integer
      id:
        type: integer
      length:
        type: integer
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: HTTP server for a service that makes http requests to 3rd-party services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Task Third Party HTTP Server
  version: "1.0"
paths:
  /task:
    post:
      consumes:
      - application/json
      description: Create a Task
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateTaskResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create Task
      tags:
      - Task
  /task/{id}:
    get:
      consumes:
      - application/json
      description: Get a Task, given the id
      parameters:
      - description: Task ID
        format: integer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetTaskResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Task
      tags:
      - Task
swagger: "2.0"